adding a plugin store so tenant admins can manage resources for their plugins. Types of resources a plugin supports will be defined by the plugin. For example, the 'chef-solo' automator type will specify that it supports 'cookbook', 'databag', and 'role' resource types. A tenant admin will then be able to upload resources for cookbooks, databags, and roles with a given name and version for identification.
Resource versions can then be activated, meaning that version
of the resource will be pushed to provisioners when a special
'sync' call is made by the admin.

Implementation includes a PluginResourceMetaStoreView that keeps
track of what resources exist and which versions are active.
There is then a PluginStore that provides input and output streams
for writing to and reading from whatever backend is used to store
the plugin resources. System admins can implement the PluginStore
class to support different storage backends, with Loom including an
in-memory backend for tests and a local file system backend, with
the idea that we would add support for distributed systems that
would work better for HA purposes.