COOPR-483 changing cluster create and delete requests to return which required provider fields are missing if they are missing.
As part of this, removed provider fields from all requests except
creates and deletes since other operations do not require
communicating with providers. Also refactored cluster service a
bit to have a separate method for adding and validating user
given provider fields instead of baking it into other methods.
This was done in order to allow returning the missing fields.